+++ manager/ManifestFile.cs
@@ -15,7 +15,7 @@ using System.Globalization;
 using System.IO;
 using System.Text;
 
-namespace CraneManager
+namespace CraneLoader
 {
     public class ScriptEntry
     {