+++ manager/TestManifest.cs
@@ -13,7 +13,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using CraneManager;
+using CraneLoader;
 
 internal static class TestManifest
 {